Refactoring Sin/Cos tables for 2D scans
